# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: HRMS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List all employees
      parameters:
        - in: query
          name: departmentId
          schema:
            type: integer
        - in: query
          name: status
          schema:
            type: string
            enum: [active, inactive, on_leave]
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: List of employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployeeBody"
      responses:
        "201":
          description: Employee created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get employee by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Employee details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    put:
      operationId: updateEmployee
      tags: [employees]
      summary: Update employee
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEmployeeBody"
      responses:
        "200":
          description: Employee updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: Delete employee
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Employee deleted
...
  /geofence:
    get:
      operationId: listGeofences
      tags: [geofence]
      summary: List all geo-fence configurations
      responses:
        "200":
          description: Geo-fence configurations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Geofence"
    post:
      operationId: createGeofence
      tags: [geofence]
      summary: Create geo-fence
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGeofenceBody"
      responses:
        "201":
          description: Geo-fence created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Geofence"
  /geofence/{id}:
    put:
      operationId: updateGeofence
      tags: [geofence]
      summary: Update geo-fence
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGeofenceBody"
      responses:
        "200":
          description: Geo-fence updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Geofence"
    delete:
      operationId: deleteGeofence
      tags: [geofence]
      summary: Delete geo-fence
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Geo-fence deleted

  # HOLIDAYS
  /holidays:
    get:
      operationId: listHolidays
      tags: [holidays]
      summary: List holidays
      parameters:
        - in: query
          name: year
          schema:
            type: integer
      responses:
        "200":
          description: Holidays
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Holiday"
    post:
      operationId: createHoliday
      tags: [holidays]
      summary: Create holiday
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHolidayBody"
      responses:
        "201":
          description: Holiday created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Holiday"
  /holidays/{id}:
    delete:
      operationId: deleteHoliday
      tags: [holidays]
      summary: Delete holiday
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Holiday deleted

  # DASHBOARD
...
components:
  schemas:
...
    Employee:
      type: object
      properties:
        id:
          type: integer
        employeeCode:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [admin, employee, manager]
        status:
          type: string
          enum: [active, inactive, on_leave]
        departmentId:
          type: integer
        departmentName:
          type: string
        designation:
          type: string
        shift:
          type: string
        joiningDate:
          type: string
          format: date
        salary:
          type: number
        avatarUrl:
          type: string
        branchId:
          type: integer
        branchName:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - employeeCode
        - firstName
        - lastName
        - email
        - role
        - status
    CreateEmployeeBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, employee, manager]
        departmentId:
          type: integer
        designation:
          type: string
        shift:
          type: string
        joiningDate:
          type: string
          format: date
        salary:
          type: number
        branchId:
          type: integer
      required:
        - firstName
        - lastName
        - email
        - password
        - role
    UpdateEmployeeBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        role:
          type: string
        departmentId:
          type: integer
        designation:
          type: string
        shift:
          type: string
        salary:
          type: number
        status:
          type: string
        branchId:
          type: integer
...
    Geofence:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        radius:
          type: number
        address:
          type: string
        branchId:
          type: integer
        isActive:
          type: boolean
      required:
        - id
        - name
        - latitude
        - longitude
        - radius
        - isActive
    CreateGeofenceBody:
      type: object
      properties:
        name:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        radius:
          type: number
        address:
          type: string
        branchId:
          type: integer
      required:
        - name
        - latitude
        - longitude
        - radius
    UpdateGeofenceBody:
      type: object
      properties:
        name:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        radius:
          type: number
        address:
          type: string
        isActive:
          type: boolean
    Holiday:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        date:
          type: string
          format: date
        type:
          type: string
          enum: [national, optional, restricted]
        description:
          type: string
      required:
        - id
        - name
        - date
        - type
    CreateHolidayBody:
      type: object
      properties:
        name:
          type: string
        date:
          type: string
          format: date
        type:
          type: string
          enum: [national, optional, restricted]
        description:
          type: string
      required:
        - name
        - date
        - type
...