# lib/api-zod/src/generated/types/createGeofenceBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HRMS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateGeofenceBody {
  name: string;
  latitude: number;
  longitude: number;
  radius: number;
  address?: string;
  branchId?: number;
}