# lib/api-zod/src/generated/types/updateEmployeeBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HRMS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateEmployeeBody {
  firstName?: string;
  lastName?: string;
  phone?: string;
  role?: string;
  departmentId?: number;
  designation?: string;
  shift?: string;
  salary?: number;
  status?: string;
  branchId?: number;
}